Specified jenkins user in withRegistry block

+++ Jenkinsfile
@@ -3,22 +3,54 @@ pipeline {
     agent any
 
     environment {
-        image_label = "flights-microservice-ab"
-        git_commit_hash = "${sh(returnStdout: true, script: 'git rev-parse --short=8 HEAD')}"
-        image = ""
+        COMMIT_HASH = sh(returnStdout: true, script: "git rev-parse --short=8 HEAD").trim()
+        AWS_REGION = sh(script:'aws configure get region', returnStdout: true).trim()
+        AWS_ACCOUNT_ID = sh(script:'aws sts get-caller-identity --query "Account" --output text', returnStdout: true).trim()
+        ECR_URI = "${AWS_ACCOUNT_ID}.dkr.ecr.${AWS_REGION}.amazonaws.com"
+        PROJECT_ID  = "AB"
+
+        image_label = "flights-microservice-${PROJECT_ID.toLowerCase()}"
+        image = null
+        packaged = false
+        built = false
     }
 
     stages {
         stage('Package') {
             steps {
-                sh "./mvnw clean package"
+                sh "./mvnw package"
+                sh "docker context use default"
+            }
+
+            post {
+                success {
+                    script {
+                        packaged = true
+                    }
+                }
             }
         }
+
+        //stage('SonarQube Analysis') {
+        //    steps {
+        //        withSonarQubeEnv('SonarQube') {
+        //            sh './mvnw org.sonarsource.scanner.maven:sonar-maven-plugin:3.7.0.1746:sonar'
+        //        }
+        //    }
+        //}
 
         stage('Build') {
             steps {
                 script {
-                    image = docker.build image_label
+                    image = docker.build(image_label)
+                }
+            }
+
+            post {
+                success {
+                    script {
+                        built = true
+                    }
                 }
             }
         }
@@ -26,18 +58,32 @@ pipeline {
         stage('Push to registry') {
             steps {
                 script {
-                    docker.withRegistry(FLIGHTS_ECR_URI_AB, 'ecr:us-west-2:ecr-creds') {
-                        image.push("$git_commit_hash")
+                    docker.withRegistry(
+                        "$ECR_URI/$image_label",
+                        "ecr:$AWS_REGION:jenkins"
+                    ) {
+                        image.push("$COMMIT_HASH")
                         image.push('latest')
                     }
                 }
             }
         }
+    }
 
-        stage('Clean up') {
-            steps {
-                sh "docker rmi $image_label"
+    post {
+        cleanup {
+            script {
+                if(packaged) {
+                    sh "./mvnw clean"
+
+                    if(built) {
+                        sh "docker rmi $image_label"
+                    }
+                }
             }
         }
     }
 }
+
+
+
